third_party/WebKit/LayoutTests/editing/execCommand/indent_at_end_of_paragraph.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable>',
                '^This paragraph should be indented.<br>|',
                'This paragraph should not be indented.<br>',
            '</div>',
        ].join(''),
        'indent',
        [
            '<div contenteditable>',
                '<blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">',
                    '^This paragraph should be indented.|',
                '</blockquote>',
                'This paragraph should not be indented.<br>',
            '</div>',
        ].join('')),
        'Indent on a selection that ends at the start of a paragraph.');
</script>
